refactoring ets + delete unused imports
