+++ lib/features/app/navigation/router.dart
@@ -7,10 +7,10 @@ import 'package:ets_api_clients/models.dart';
 // Project imports:
 import 'package:notredame/features/app/navigation/router_paths.dart';
 import 'package:notredame/constants/update_code.dart';
-import 'package:notredame/features/ets/author/author_view.dart';
+import 'package:notredame/features/ets/events/author/author_view.dart';
 import 'package:notredame/features/ets/ets_view.dart';
-import 'package:notredame/features/ets/news/news-details/news_details_view.dart';
-import 'package:notredame/features/ets/news/news_view.dart';
+import 'package:notredame/features/ets/events/news/news-details/news_details_view.dart';
+import 'package:notredame/features/ets/events/news/news_view.dart';
 import 'package:notredame/features/ets/quick-link/models/quick_link.dart';
 import 'package:notredame/features/more/about/about_view.dart';
 import 'package:notredame/features/more/settings/choose_language_view.dart';
@@ -26,7 +26,7 @@ import 'package:notredame/features/app/error/outage/outage_view.dart';
 import 'package:notredame/features/ets/quick-link/quick_links_view.dart';
 import 'package:notredame/features/schedule/schedule_default/schedule_default_view.dart';
 import 'package:notredame/features/schedule/schedule_view.dart';
-import 'package:notredame/features/ets/security-info/security_view.dart';
+import 'package:notredame/features/ets/web-link/security-info/security_view.dart';
 import 'package:notredame/features/more/settings/settings_view.dart';
 import 'package:notredame/features/app/startup/startup_view.dart';
 import 'package:notredame/features/student/student_view.dart';
