+++ lib/features/ets/events/author/author_view.dart
@@ -6,18 +6,18 @@ import 'package:flutter/material.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 import 'package:font_awesome_flutter/font_awesome_flutter.dart';
 import 'package:infinite_scroll_pagination/infinite_scroll_pagination.dart';
-import 'package:notredame/features/ets/social/models/social_link.dart';
+import 'package:notredame/features/ets/events/social/models/social_link.dart';
 import 'package:stacked/stacked.dart';
 
 // Project imports:
 import 'package:notredame/utils/utils.dart';
-import 'package:notredame/features/ets/author/author_viewmodel.dart';
+import 'package:notredame/features/ets/events/author/author_viewmodel.dart';
 import 'package:notredame/utils/app_theme.dart';
-import 'package:notredame/features/ets/author/author_info_skeleton.dart';
+import 'package:notredame/features/ets/events/author/author_info_skeleton.dart';
 import 'package:notredame/features/app/widgets/base_scaffold.dart';
-import 'package:notredame/features/ets/news/news_card.dart';
-import 'package:notredame/features/ets/news/news_card_skeleton.dart';
-import 'package:notredame/features/ets/social/social_links_card.dart';
+import 'package:notredame/features/ets/events/news/news_card.dart';
+import 'package:notredame/features/ets/events/news/news_card_skeleton.dart';
+import 'package:notredame/features/ets/events/social/social_links_card.dart';
 
 class AuthorView extends StatefulWidget {
   final String authorId;
