+++ lib/features/ets/events/news/news-details/news_details_view.dart
@@ -8,7 +8,7 @@ import 'package:flutter_markdown/flutter_markdown.dart';
 import 'package:flutter_svg/svg.dart';
 import 'package:intl/intl.dart';
 import 'package:notredame/features/app/analytics/remote_config_service.dart';
-import 'package:notredame/features/ets/report_news.dart';
+import 'package:notredame/features/ets/events/report_news.dart';
 import 'package:notredame/utils/utils.dart';
 import 'package:notredame/features/schedule/calendar_selection_viewmodel.dart';
 import 'package:notredame/features/schedule/widgets/calendar_selector.dart';
@@ -19,7 +19,7 @@ import 'package:stacked/stacked.dart';
 import 'package:notredame/features/app/navigation/router_paths.dart';
 import 'package:notredame/features/app/analytics/analytics_service.dart';
 import 'package:notredame/features/app/navigation/navigation_service.dart';
-import 'package:notredame/features/ets/news/news-details/news_details_viewmodel.dart';
+import 'package:notredame/features/ets/events/news/news-details/news_details_viewmodel.dart';
 import 'package:notredame/utils/locator.dart';
 import 'package:notredame/utils/app_theme.dart';
 import 'package:notredame/features/app/widgets/base_scaffold.dart';
