+++ lib/features/ets/events/news/news_card_skeleton.dart
