+++ lib/features/ets/events/news/news_view.dart
@@ -10,9 +10,9 @@ import 'package:stacked/stacked.dart';
 import 'package:infinite_scroll_pagination/infinite_scroll_pagination.dart';
 
 // Project imports:
-import 'package:notredame/features/ets/news/news_viewmodel.dart';
-import 'package:notredame/features/ets/news/news_card.dart';
-import 'package:notredame/features/ets/news/news_card_skeleton.dart';
+import 'package:notredame/features/ets/events/news/news_viewmodel.dart';
+import 'package:notredame/features/ets/events/news/news_card.dart';
+import 'package:notredame/features/ets/events/news/news_card_skeleton.dart';
 
 class NewsView extends StatefulWidget {
   @override
