+++ lib/features/ets/events/report_news.dart
@@ -8,9 +8,9 @@ import 'package:notredame/utils/app_theme.dart';
 import 'package:stacked/stacked.dart';
 
 // Project imports:
-import 'package:notredame/features/ets/report-news/report_news.dart';
+import 'package:notredame/features/ets/events/report-news/report_news.dart';
 import 'package:notredame/utils/utils.dart';
-import 'package:notredame/features/ets/report-news/report_news_viewmodel.dart';
+import 'package:notredame/features/ets/events/report-news/report_news_viewmodel.dart';
 
 class ReportNews extends StatefulWidget {
   final bool showHandle;
