+++ test/ui/views/author_view_test.dart
@@ -17,9 +17,9 @@ import 'package:notredame/features/app/analytics/analytics_service.dart';
 import 'package:notredame/features/app/integration/launch_url_service.dart';
 import 'package:notredame/features/app/navigation/navigation_service.dart';
 import 'package:notredame/features/app/integration/networking_service.dart';
-import 'package:notredame/features/ets/author/author_view.dart';
-import 'package:notredame/features/ets/news/news_card.dart';
-import 'package:notredame/features/ets/social/social_links_card.dart';
+import 'package:notredame/features/ets/events/author/author_view.dart';
+import 'package:notredame/features/ets/events/news/news_card.dart';
+import 'package:notredame/features/ets/events/social/social_links_card.dart';
 import '../../helpers.dart';
 import '../../mock/managers/author_repository_mock.dart';
 import '../../mock/managers/news_repository_mock.dart';
