+++ test/ui/views/news_details_view_test.dart
@@ -13,7 +13,7 @@ import 'package:notredame/features/app/analytics/analytics_service.dart';
 import 'package:notredame/features/app/navigation/navigation_service.dart';
 import 'package:notredame/features/app/integration/networking_service.dart';
 import 'package:notredame/features/app/analytics/remote_config_service.dart';
-import 'package:notredame/features/ets/news/news-details/news_details_view.dart';
+import 'package:notredame/features/ets/events/news/news-details/news_details_view.dart';
 import '../../helpers.dart';
 
 void main() {
