+++ test/ui/views/news_view_test.dart
@@ -13,8 +13,8 @@ import 'package:notredame/features/app/repository/news_repository.dart';
 import 'package:notredame/features/more/settings/settings_manager.dart';
 import 'package:notredame/features/app/navigation/navigation_service.dart';
 import 'package:notredame/features/app/integration/networking_service.dart';
-import 'package:notredame/features/ets/news/news_view.dart';
-import 'package:notredame/features/ets/news/news_card.dart';
+import 'package:notredame/features/ets/events/news/news_view.dart';
+import 'package:notredame/features/ets/events/news/news_card.dart';
 import '../../helpers.dart';
 import '../../mock/managers/news_repository_mock.dart';
