+++ test/ui/views/security_view_test.dart
@@ -7,8 +7,8 @@ import 'package:flutter_test/flutter_test.dart';
 import 'package:google_maps_flutter/google_maps_flutter.dart';
 
 // Project imports:
-import 'package:notredame/features/ets/security-info/models/emergency_procedures.dart';
-import 'package:notredame/features/ets/security-info/security_view.dart';
+import 'package:notredame/features/ets/web-link/security-info/models/emergency_procedures.dart';
+import 'package:notredame/features/ets/web-link/security-info/security_view.dart';
 import '../../helpers.dart';
 
 void main() {
