+++ test/ui/widgets/news_card_test.dart
@@ -6,7 +6,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 
 // Project imports:
-import 'package:notredame/features/ets/news/news_card.dart';
+import 'package:notredame/features/ets/events/news/news_card.dart';
 import '../../helpers.dart';
 
 void main() {
