+++ test/ui/widgets/news_skeleton_test.dart
@@ -3,7 +3,7 @@ import 'package:flutter_test/flutter_test.dart';
 import 'package:shimmer/shimmer.dart';
 
 // Project imports:
-import 'package:notredame/features/ets/news/news_card_skeleton.dart';
+import 'package:notredame/features/ets/events/news/news_card_skeleton.dart';
 import '../../helpers.dart';
 
 void main() {
