+++ test/viewmodels/news_details_viewmodel_test.dart
@@ -3,7 +3,7 @@ import 'package:ets_api_clients/models.dart';
 import 'package:flutter_test/flutter_test.dart';
 
 // Project imports:
-import 'package:notredame/features/ets/news/news-details/news_details_viewmodel.dart';
+import 'package:notredame/features/ets/events/news/news-details/news_details_viewmodel.dart';
 import 'package:notredame/utils/app_theme.dart';
 
 void main() {
