+++ test/viewmodels/news_viewmodel_test.dart
@@ -6,7 +6,7 @@ import 'package:mockito/mockito.dart';
 // Project imports:
 import 'package:notredame/features/app/repository/news_repository.dart';
 import 'package:notredame/features/more/settings/settings_manager.dart';
-import 'package:notredame/features/ets/news/news_viewmodel.dart';
+import 'package:notredame/features/ets/events/news/news_viewmodel.dart';
 import 'package:notredame/utils/locator.dart';
 import '../helpers.dart';
 import '../mock/managers/news_repository_mock.dart';
